Add gate for authorized user

+++ app/Http/Controllers/QuestionsController.php
@@ -66,7 +66,12 @@ class QuestionsController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Question $question)
-    {
+    {   
+
+        if(\Gate::denies('update-questions', $question)) {
+            abort(403, 'Access denied!');
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -79,6 +84,11 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+
+        if(\Gate::denies('update-questions', $question)) {
+            abort(403, 'Access denied!');
+        }
+
         $question->update($request->all());
         
         return redirect('/questions')->with('success', 'Your Question has been updated');
@@ -92,6 +102,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+
+        if(\Gate::denies('delete-questions', $question)) {
+            abort(403, 'Access denied!');
+        }
+
         $question->delete();
         
         return redirect('/questions')->with('success', 'Your Question as been deleted');
